refact: create lib/security folder
